requirements.txt
matplotlib
nibabel
numpy
Pillow
PyYAML
scikit_image
skimage
torch
torchvision
